examples/pybind11_requirements.txt
bottle==0.12.23
numpy #==1.24.0
Pillow #==10.1.0
#Pillow-SIMD==9.0.0.post1
